move TopicColor to viewmodel
